.platform/requirements.txt
PyYAML==5.4.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2